requirements.txt
mysql-connector
pymysql
pika
connexion[swagger-ui]
flask-sqlalchemy